+++ clang-tidy/nodecpp/RawPtrExprCheck.cpp
@@ -1,4 +1,4 @@
-//===--- NakedStructCheck.cpp - clang-tidy---------------------------------===//
+//===--- RawPtrExprCheck.cpp - clang-tidy----------------------------------===//
 //
 //                     The LLVM Compiler Infrastructure
 //
@@ -7,7 +7,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "NakedStructCheck.h"
+#include "RawPtrExprCheck.h"
 #include "NakedPtrHelper.h"
 #include "clang/AST/ASTContext.h"
 #include "clang/ASTMatchers/ASTMatchFinder.h"
@@ -18,7 +18,7 @@ namespace clang {
 namespace tidy {
 namespace nodecpp {
 
-void NakedStructCheck::registerMatchers(MatchFinder *Finder) {
+void RawPtrExprCheck::registerMatchers(MatchFinder *Finder) {
   Finder->addMatcher(expr(hasType(pointerType())).bind("expr"),
       this);
 
@@ -26,42 +26,42 @@ void NakedStructCheck::registerMatchers(MatchFinder *Finder) {
       this);
 }
 
-void NakedStructCheck::check(const MatchFinder::MatchResult &Result) {
+void RawPtrExprCheck::check(const MatchFinder::MatchResult &Result) {
   auto expr = Result.Nodes.getNodeAs<Expr>("expr");
 
   //first ignore implicits and parens
 
   if(isa<ExprWithCleanups>(expr))
     return;
-  if(isa<MaterializeTemporaryExpr>(expr))
+  else if(isa<MaterializeTemporaryExpr>(expr))
     return;
-  if(isa<CXXBindTemporaryExpr>(expr))
+  else if(isa<CXXBindTemporaryExpr>(expr))
     return;
-  if(isa<ImplicitCastExpr>(expr))
+  else if(isa<ImplicitCastExpr>(expr))
     return;
-  if(isa<ParenExpr>(expr))
+  else if(isa<ParenExpr>(expr))
     return;
-  if(isa<CXXDefaultArgExpr>(expr))
+  else if(isa<CXXDefaultArgExpr>(expr))
     return;
 
   // now allow some explicits
-  if(isa<CXXThisExpr>(expr))
+  else if(isa<CXXThisExpr>(expr))
     return;
-  if(isa<CallExpr>(expr))
+  else if(isa<CallExpr>(expr))
     return;
-  if(isa<CXXNullPtrLiteralExpr>(expr))
+  else if(isa<CXXNullPtrLiteralExpr>(expr))
     return;
-  if(isa<CXXDynamicCastExpr>(expr))
+  else if(isa<CXXDynamicCastExpr>(expr))
     return;
 
-  if(auto unOp = dyn_cast<UnaryOperator>(expr)) {
+  else if(auto unOp = dyn_cast<UnaryOperator>(expr)) {
     if(unOp->getOpcode() == UnaryOperatorKind::UO_AddrOf) {
-      diag(expr->getExprLoc(), "(S1) address-of operator found here");
+      //this is ok
       return;
     }
   }
 
-  if(auto binOp = dyn_cast<BinaryOperator>(expr)) {
+  else if(auto binOp = dyn_cast<BinaryOperator>(expr)) {
     if(binOp->getOpcode() == BinaryOperatorKind::BO_Assign) {
       //this is ok
       return;
@@ -70,30 +70,30 @@ void NakedStructCheck::check(const MatchFinder::MatchResult &Result) {
 
   // this is an error,
   // find the best error message
-  if(isa<DeclRefExpr>(expr)) {
+  else if(isa<DeclRefExpr>(expr)) {
       // don't report it here
       return;
   }
-  if(isa<CXXNewExpr>(expr)) {
+  else if(isa<CXXNewExpr>(expr)) {
       // don't report here
       return;
   }
-  if(isa<CXXDeleteExpr>(expr)) {
+  else if(isa<CXXDeleteExpr>(expr)) {
       // don't report here
       return;
   }
 
-  if(isa<CXXStaticCastExpr>(expr)) {
+  else if(isa<CXXStaticCastExpr>(expr)) {
     diag(expr->getExprLoc(), "(S1.1) static_cast not allowed");
     return;
   }
 
-  if(isa<CXXReinterpretCastExpr>(expr)) {
+  else if(isa<CXXReinterpretCastExpr>(expr)) {
     diag(expr->getExprLoc(), "(S1.1) reinterpret_cast not allowed");
     return;
   }
 
-  if(isa<CStyleCastExpr>(expr)) {
+  else if(isa<CStyleCastExpr>(expr)) {
     diag(expr->getExprLoc(), "(S1.1) C style cast not allowed");
     return;
   }