+++ clang-tidy/nodecpp/RawPtrExprCheck.h
@@ -1,4 +1,4 @@
-//===--- NakedStructCheck.h - clang-tidy-------------------------*- C++ -*-===//
+//===--- RawPtrExprCheck.h - clang-tidy-------------------------*- C++ -*-===//
 //
 //                     The LLVM Compiler Infrastructure
 //
@@ -19,10 +19,10 @@ namespace nodecpp {
 /// FIXME: Write a short description.
 ///
 /// For the user-facing documentation see:
-/// http://clang.llvm.org/extra/clang-tidy/checks/nodecpp-naked-struct.html
-class NakedStructCheck : public ClangTidyCheck {
+/// http://clang.llvm.org/extra/clang-tidy/checks/nodecpp-raw-ptr-expr.html
+class RawPtrExprCheck : public ClangTidyCheck {
 public:
-  NakedStructCheck(StringRef Name, ClangTidyContext *Context)
+  RawPtrExprCheck(StringRef Name, ClangTidyContext *Context)
       : ClangTidyCheck(Name, Context) {}
   void registerMatchers(ast_matchers::MatchFinder *Finder) override;
   void check(const ast_matchers::MatchFinder::MatchResult &Result) override;